refactor: handlerAdapter 구현체 패키지 위치 이동
