+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/ControllerHandlerAdapter.java
@@ -1,9 +1,9 @@
-package com.interface21.webmvc.servlet.mvc;
+package com.interface21.webmvc.servlet.mvc.asis;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.HandlerAdapter;
 import com.interface21.webmvc.servlet.view.JspView;
 
 public class ControllerHandlerAdapter implements HandlerAdapter {
